Move some of the plumbing for tests into separate files
No functional changes
