+++ pkg/controller/bundlec_test/BUILD.bazel
@@ -17,6 +17,7 @@ go_test(
         "no_actions_for_blocked_resources_test.go",
         "no_deletions_while_in_progress_test.go",
         "not_marked_crd_ignored_test.go",
+        "plugin_error_propagated_test.go",
         "plugin_schema_invalid_test.go",
         "plugin_spec_processed_test.go",
         "processing_continues_after_error_test.go",
@@ -25,6 +26,8 @@ go_test(
         "secret_keys_not_merged_test.go",
         "service_instance_schema_invalid_test.go",
         "two_resources_same_name_test.go",
+        "zz_objects_for_test.go",
+        "zz_plugins_for_test.go",
         "zz_plumbing_for_test.go",
     ],
     race = "on",
