+++ pkg/controller/bundlec_test/zz_plumbing_for_test.go
@@ -13,8 +13,6 @@ import (
 	"github.com/atlassian/ctrl"
 	"github.com/atlassian/ctrl/process"
 	"github.com/atlassian/smith/cmd/smith/app"
-	"github.com/atlassian/smith/examples/sleeper"
-	sleeper_v1 "github.com/atlassian/smith/examples/sleeper/pkg/apis/sleeper/v1"
 	smith_v1 "github.com/atlassian/smith/pkg/apis/smith/v1"
 	smithFake "github.com/atlassian/smith/pkg/client/clientset_generated/clientset/fake"
 	"github.com/atlassian/smith/pkg/client/smart"
@@ -24,13 +22,11 @@ import (
 	sc_v1b1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	scClientset "github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset"
 	scFake "github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset/fake"
-	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest"
-	core_v1 "k8s.io/api/core/v1"
 	apiext_v1b1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	apiExtFake "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/fake"
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -485,17 +481,6 @@ func (f *fakeActionHandler) ServeHTTP(response http.ResponseWriter, request *htt
 	response.Write(fakeResp.content)
 }
 
-func SleeperCrdWithStatus() *apiext_v1b1.CustomResourceDefinition {
-	crd := sleeper.Crd()
-	crd.Status = apiext_v1b1.CustomResourceDefinitionStatus{
-		Conditions: []apiext_v1b1.CustomResourceDefinitionCondition{
-			{Type: apiext_v1b1.Established, Status: apiext_v1b1.ConditionTrue},
-			{Type: apiext_v1b1.NamesAccepted, Status: apiext_v1b1.ConditionTrue},
-		},
-	}
-	return crd
-}
-
 func convertBundleResourcesToUnstrucutred(t *testing.T, bundle *smith_v1.Bundle, serviceCatalog bool) {
 	// Convert all typed objects into unstructured ones
 	for i, res := range bundle.Spec.Resources {
@@ -504,388 +489,5 @@ func convertBundleResourcesToUnstrucutred(t *testing.T, bundle *smith_v1.Bundle,
 			require.NoError(t, err)
 			bundle.Spec.Resources[i].Spec.Object = resUnstr
 		}
-	}
-}
-
-type testingPlugin interface {
-	plugin.Plugin
-	WasInvoked() bool
-}
-
-func configMapWithDependenciesPlugin(expectBinding, expectSleeper bool) func(t *testing.T) testingPlugin {
-	return func(t *testing.T) testingPlugin {
-		return &configMapWithDeps{
-			t:             t,
-			expectBinding: expectBinding,
-			expectSleeper: expectSleeper,
-		}
-	}
-}
-
-type configMapWithDeps struct {
-	t             *testing.T
-	expectBinding bool
-	expectSleeper bool
-	wasInvoked    bool
-}
-
-func (p *configMapWithDeps) WasInvoked() bool {
-	return p.wasInvoked
-}
-
-func (p *configMapWithDeps) Describe() *plugin.Description {
-	return &plugin.Description{
-		Name: pluginConfigMapWithDeps,
-		GVK:  core_v1.SchemeGroupVersion.WithKind("ConfigMap"),
-		SpecSchema: []byte(`{
-			"type": "object",
-			"properties": {
-				"p1": {
-					"type": "string"
-				},
-				"p2": {
-					"type": "string"
-				}
-			}
-		}`),
-	}
-}
-
-func (p *configMapWithDeps) Process(pluginSpec map[string]interface{}, context *plugin.Context) (*plugin.ProcessResult, error) {
-	p.wasInvoked = true
-	failed := p.t.Failed()
-
-	assert.Equal(p.t, testNamespace, context.Namespace)
-
-	actualShouldExist, _ := pluginSpec["actualShouldExist"].(bool)
-	delete(pluginSpec, "actualShouldExist")
-	assert.Equal(p.t, map[string]interface{}{"p1": "v1", "p2": sb1}, pluginSpec)
-
-	if actualShouldExist {
-		assert.IsType(p.t, &core_v1.ConfigMap{}, context.Actual)
-	} else {
-		assert.Nil(p.t, context.Actual)
-	}
-
-	bindingDep, ok := context.Dependencies[resSb1]
-	if p.expectBinding && assert.True(p.t, ok) {
-		// Actual
-		if assert.IsType(p.t, &sc_v1b1.ServiceBinding{}, bindingDep.Actual) {
-			b := bindingDep.Actual.(*sc_v1b1.ServiceBinding)
-			assert.Equal(p.t, sb1, b.Name)
-			assert.Equal(p.t, testNamespace, b.Namespace)
-			assert.Equal(p.t, sc_v1b1.SchemeGroupVersion.WithKind("ServiceBinding"), b.GroupVersionKind())
-		}
-		// Outputs
-		if assert.Len(p.t, bindingDep.Outputs, 1) {
-			secret := bindingDep.Outputs[0]
-			if assert.IsType(p.t, &core_v1.Secret{}, secret) {
-				s := secret.(*core_v1.Secret)
-				assert.Equal(p.t, s1, s.Name)
-				assert.Equal(p.t, testNamespace, s.Namespace)
-				assert.Equal(p.t, core_v1.SchemeGroupVersion.WithKind("Secret"), s.GroupVersionKind())
-			}
-		}
-		// Aux
-		if assert.Len(p.t, bindingDep.Auxiliary, 1) {
-			svcInst := bindingDep.Auxiliary[0]
-			if assert.IsType(p.t, &sc_v1b1.ServiceInstance{}, svcInst) {
-				inst := svcInst.(*sc_v1b1.ServiceInstance)
-				assert.Equal(p.t, si1, inst.Name)
-				assert.Equal(p.t, testNamespace, inst.Namespace)
-				assert.Equal(p.t, sc_v1b1.SchemeGroupVersion.WithKind("ServiceInstance"), inst.GroupVersionKind())
-			}
-		}
-	}
-	sleeperDep, ok := context.Dependencies[resSleeper1]
-	if p.expectSleeper && assert.True(p.t, ok) {
-		// Actual
-		if assert.IsType(p.t, &unstructured.Unstructured{}, sleeperDep.Actual) {
-			s := sleeperDep.Actual.(*unstructured.Unstructured)
-			assert.Equal(p.t, sleeper1, s.GetName())
-			assert.Equal(p.t, testNamespace, s.GetNamespace())
-			assert.Equal(p.t, sleeper_v1.SleeperGVK, s.GroupVersionKind())
-		}
-		// Outputs
-		assert.Empty(p.t, sleeperDep.Outputs)
-		// Aux
-		assert.Empty(p.t, sleeperDep.Auxiliary)
-	}
-
-	if !failed && p.t.Failed() { // one of the assertions failed and it was the first failure in the test
-		return nil, errors.New("plugin failed BOOM!")
-	}
-
-	return &plugin.ProcessResult{
-		Object: &core_v1.ConfigMap{
-			TypeMeta: meta_v1.TypeMeta{
-				Kind:       "ConfigMap",
-				APIVersion: core_v1.SchemeGroupVersion.String(),
-			},
-		},
-	}, nil
-}
-
-func simpleConfigMapPlugin(t *testing.T) testingPlugin {
-	return &simpleConfigMap{
-		t: t,
-	}
-}
-
-type simpleConfigMap struct {
-	t          *testing.T
-	wasInvoked bool
-}
-
-func (p *simpleConfigMap) WasInvoked() bool {
-	return p.wasInvoked
-}
-
-func (p *simpleConfigMap) Describe() *plugin.Description {
-	return &plugin.Description{
-		Name: pluginSimpleConfigMap,
-		GVK:  core_v1.SchemeGroupVersion.WithKind("ConfigMap"),
-	}
-}
-
-func (p *simpleConfigMap) Process(pluginSpec map[string]interface{}, context *plugin.Context) (*plugin.ProcessResult, error) {
-	p.wasInvoked = true
-	failed := p.t.Failed()
-
-	assert.Equal(p.t, testNamespace, context.Namespace)
-
-	actualShouldExist, _ := pluginSpec["actualShouldExist"].(bool)
-
-	if actualShouldExist {
-		assert.IsType(p.t, &core_v1.ConfigMap{}, context.Actual)
-	} else {
-		assert.Nil(p.t, context.Actual)
-	}
-
-	if !failed && p.t.Failed() { // one of the assertions failed and it was the first failure in the test
-		return nil, errors.New("plugin failed BOOM!")
-	}
-
-	return &plugin.ProcessResult{
-		Object: &core_v1.ConfigMap{
-			TypeMeta: meta_v1.TypeMeta{
-				Kind:       "ConfigMap",
-				APIVersion: core_v1.SchemeGroupVersion.String(),
-			},
-		},
-	}, nil
-}
-
-func serviceInstance(ready, inProgress, error bool) *sc_v1b1.ServiceInstance {
-	tr := true
-	var status sc_v1b1.ServiceInstanceStatus
-	if ready {
-		status.Conditions = append(status.Conditions, sc_v1b1.ServiceInstanceCondition{
-			Type:   sc_v1b1.ServiceInstanceConditionReady,
-			Status: sc_v1b1.ConditionTrue,
-		})
-	}
-	if inProgress {
-		status.Conditions = append(status.Conditions, sc_v1b1.ServiceInstanceCondition{
-			Type:    sc_v1b1.ServiceInstanceConditionReady,
-			Status:  sc_v1b1.ConditionFalse,
-			Reason:  "WorkingOnIt",
-			Message: "Doing something",
-		})
-	}
-	if error {
-		status.Conditions = append(status.Conditions, sc_v1b1.ServiceInstanceCondition{
-			Type:    sc_v1b1.ServiceInstanceConditionFailed,
-			Status:  sc_v1b1.ConditionTrue,
-			Reason:  "BlaBla",
-			Message: "Oh no!",
-		})
-	}
-	return &sc_v1b1.ServiceInstance{
-		TypeMeta: meta_v1.TypeMeta{
-			Kind:       "ServiceInstance",
-			APIVersion: sc_v1b1.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: meta_v1.ObjectMeta{
-			Name:      si1,
-			Namespace: testNamespace,
-			UID:       si1uid,
-			OwnerReferences: []meta_v1.OwnerReference{
-				{
-					APIVersion:         smith_v1.BundleResourceGroupVersion,
-					Kind:               smith_v1.BundleResourceKind,
-					Name:               bundle1,
-					UID:                bundle1uid,
-					Controller:         &tr,
-					BlockOwnerDeletion: &tr,
-				},
-			},
-		},
-		Status: status,
-		Spec:   serviceInstanceSpec,
-	}
-}
-
-func serviceBinding(ready, inProgress, error bool) *sc_v1b1.ServiceBinding {
-	tr := true
-	var status sc_v1b1.ServiceBindingStatus
-	if ready {
-		status.Conditions = append(status.Conditions, sc_v1b1.ServiceBindingCondition{
-			Type:   sc_v1b1.ServiceBindingConditionReady,
-			Status: sc_v1b1.ConditionTrue,
-		})
-	}
-	if inProgress {
-		status.Conditions = append(status.Conditions, sc_v1b1.ServiceBindingCondition{
-			Type:    sc_v1b1.ServiceBindingConditionReady,
-			Status:  sc_v1b1.ConditionFalse,
-			Reason:  "WorkingOnIt",
-			Message: "Doing something",
-		})
-	}
-	if error {
-		status.Conditions = append(status.Conditions, sc_v1b1.ServiceBindingCondition{
-			Type:    sc_v1b1.ServiceBindingConditionFailed,
-			Status:  sc_v1b1.ConditionTrue,
-			Reason:  "BlaBla",
-			Message: "Oh no!",
-		})
-	}
-	return &sc_v1b1.ServiceBinding{
-		TypeMeta: meta_v1.TypeMeta{
-			Kind:       "ServiceBinding",
-			APIVersion: sc_v1b1.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: meta_v1.ObjectMeta{
-			Name:      sb1,
-			Namespace: testNamespace,
-			UID:       sb1uid,
-			OwnerReferences: []meta_v1.OwnerReference{
-				{
-					APIVersion:         smith_v1.BundleResourceGroupVersion,
-					Kind:               smith_v1.BundleResourceKind,
-					Name:               bundle1,
-					UID:                bundle1uid,
-					Controller:         &tr,
-					BlockOwnerDeletion: &tr,
-				},
-				{
-					APIVersion:         sc_v1b1.SchemeGroupVersion.String(),
-					Kind:               "ServiceInstance",
-					Name:               si1,
-					UID:                si1uid,
-					BlockOwnerDeletion: &tr,
-				},
-			},
-		},
-		Spec: sc_v1b1.ServiceBindingSpec{
-			ServiceInstanceRef: sc_v1b1.LocalObjectReference{
-				Name: si1,
-			},
-			SecretName: s1,
-		},
-		Status: status,
-	}
-}
-
-func configMapNeedsUpdate() *core_v1.ConfigMap {
-	tr := true
-	return &core_v1.ConfigMap{
-		TypeMeta: meta_v1.TypeMeta{
-			Kind:       "ConfigMap",
-			APIVersion: core_v1.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: meta_v1.ObjectMeta{
-			Name:      mapNeedsAnUpdate,
-			Namespace: testNamespace,
-			UID:       mapNeedsAnUpdateUid,
-			OwnerReferences: []meta_v1.OwnerReference{
-				{
-					APIVersion: smith_v1.BundleResourceGroupVersion,
-					Kind:       smith_v1.BundleResourceKind,
-					Name:       bundle1,
-					UID:        bundle1uid,
-					Controller: &tr,
-					//BlockOwnerDeletion is missing - needs an update
-				},
-			},
-		},
-		Data: map[string]string{
-			"delete": "this key",
-		},
-	}
-}
-
-func configMapNeedsUpdateResponse(bundleName string, bundleUid types.UID) fakeResponse {
-	return fakeResponse{
-		statusCode: http.StatusOK,
-		content: []byte(`{
-							"apiVersion": "v1",
-							"kind": "ConfigMap",
-							"metadata": {
-								"name": "` + mapNeedsAnUpdate + `",
-								"namespace": "` + testNamespace + `",
-								"uid": "` + string(mapNeedsAnUpdateUid) + `",
-								"ownerReferences": [{
-									"apiVersion": "` + smith_v1.BundleResourceGroupVersion + `",
-									"kind": "` + smith_v1.BundleResourceKind + `",
-									"name": "` + bundleName + `",
-									"uid": "` + string(bundleUid) + `",
-									"controller": true,
-									"blockOwnerDeletion": true
-								}] }
-							}`),
-	}
-}
-
-func configMapNeedsDelete() *core_v1.ConfigMap {
-	tr := true
-	return &core_v1.ConfigMap{
-		TypeMeta: meta_v1.TypeMeta{
-			Kind:       "ConfigMap",
-			APIVersion: core_v1.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: meta_v1.ObjectMeta{
-			Name:      mapNeedsDelete,
-			Namespace: testNamespace,
-			UID:       mapNeedsDeleteUid,
-			OwnerReferences: []meta_v1.OwnerReference{
-				{
-					APIVersion:         smith_v1.BundleResourceGroupVersion,
-					Kind:               smith_v1.BundleResourceKind,
-					Name:               bundle1,
-					UID:                bundle1uid,
-					Controller:         &tr,
-					BlockOwnerDeletion: &tr,
-				},
-			},
-		},
-	}
-}
-
-func configMapMarkedForDeletion() *core_v1.ConfigMap {
-	tr := true
-	now := meta_v1.Now()
-	return &core_v1.ConfigMap{
-		TypeMeta: meta_v1.TypeMeta{
-			Kind:       "ConfigMap",
-			APIVersion: core_v1.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: meta_v1.ObjectMeta{
-			Name:              mapMarkedForDeletion,
-			Namespace:         testNamespace,
-			UID:               mapMarkedForDeletionUid,
-			DeletionTimestamp: &now,
-			OwnerReferences: []meta_v1.OwnerReference{
-				{
-					APIVersion:         smith_v1.BundleResourceGroupVersion,
-					Kind:               smith_v1.BundleResourceKind,
-					Name:               bundle1,
-					UID:                bundle1uid,
-					Controller:         &tr,
-					BlockOwnerDeletion: &tr,
-				},
-			},
-		},
 	}
 }
